+++ .gitignore
@@ -6,3 +6,4 @@ db-config.json
 build
 coverage
 dist
+build_extension
