--- client/build_extension/manifest.json
@@ -1,13 +1,0 @@
-{
-	"short_name": "CasperDash",
-	"name": "CasperDash",
-	"version": "1.0.0",
-	"manifest_version": 2,
-	"icons": {
-		"512": "512.png"
-	},
-	"browser_action": {
-		"default_icon": "512.png",
-		"default_popup": "popup.html"
-	}
-}
