--- client/build_extension/popup.html
@@ -1,43 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<!-- <link rel="icon" href="%PUBLIC_URL%/favicon.ico" /> -->
-		<meta charset="UTF-8" />
-		<meta http-equiv="X-UA-Compatible" content="IE=edge" />
-		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-
-		<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.5.0/font/bootstrap-icons.css" />
-		<!--
-      manifest.json provides metadata used when your web app is installed on a
-      user's mobile device or desktop. See https://developers.google.com/web/fundamentals/web-app-manifest/
-    -->
-		<!-- <link rel="manifest" href="%PUBLIC_URL%/manifest.json" /> -->
-		<!--
-      Notice the use of %PUBLIC_URL% in the tags above.
-      It will be replaced with the URL of the `public` folder during the build.
-      Only files inside the `public` folder can be referenced from the HTML.
-
-      Unlike "/favicon.ico" or "favicon.ico", "%PUBLIC_URL%/favicon.ico" will
-      work correctly both with client-side routing and a non-root public URL.
-      Learn how to configure a non-root public URL by running `npm run build`.
-    -->
-		<title>Casper Dash</title>
-
-		<!-- custom style -->
-		<!-- <link rel="stylesheet" href="%PUBLIC_URL%/assets/css/style.css?ver=1.3.0" />
-		<link rel="stylesheet" href="%PUBLIC_URL%/assets/css/light-theme.css?ver=1.3.0" /> -->
-	<script defer src="static/js/main.6953a719.js"></script><link href="assets/css/main.css" rel="stylesheet"></head>
-	<body>
-		<noscript>You need to enable JavaScript to run this app.</noscript>
-		<div id="root"></div>
-		<!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
-
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.
-
-      To begin the development, run `npm start` or `yarn start`.
-      To create a production bundle, use `npm run build` or `yarn build`.
-    --></body>
-</html>
