+++ client/package.json
@@ -47,10 +47,10 @@
 		]
 	},
 	"scripts": {
-		"dev": "webpack-dev-server --mode development --history-api-fallback",
+		"dev": "webpack-dev-server --mode development",
 		"dev-extension": "webpack --mode development --env=type=extension --watch",
-		"start": "webpack-dev-server  --history-api-fallback",
-		"start-testnet": "REACT_APP_API_ROOT=https://testnet-api.casperdash.io webpack-dev-server --history-api-fallback",
+		"start": "webpack-dev-server --mode development",
+		"start-testnet": "REACT_APP_API_ROOT=https://testnet-api.casperdash.io webpack-dev-server --mode development",
 		"build": "webpack --mode production  && react-snapshot",
 		"build-testnet": "REACT_APP_API_ROOT=https://testnet-api.casperdash.io webpack --mode production  && react-snapshot",
 		"build-extension": "webpack --mode production --env=type=extension",
@@ -98,6 +98,7 @@
 		"webpack": "^5.65.0",
 		"webpack-bundle-analyzer": "^4.5.0",
 		"webpack-cli": "^4.9.1",
-		"webpack-dev-server": "^4.6.0"
+		"webpack-dev-server": "^4.6.0",
+		"webpack-merge": "^5.8.0"
 	}
 }
