--- client/src/extension/App.css
@@ -1,4 +1,0 @@
-#root {
-	width: 357px;
-	height: 600px;
-}
