+++ client/src/index.js
@@ -1,6 +1,5 @@
 import React from 'react';
 import { render } from 'react-snapshot';
-import WebExtension from './extension/App';
-// import App from './App';
+import App from './App';
 
-render(<WebExtension />, document.getElementById('root'));
+render(<App />, document.getElementById('root'));
