+++ client/src/web-extension/App.css
@@ -1,0 +1,4 @@
+#root {
+	width: 357px;
+	height: 600px;
+}
