+++ client/src/web-extension/index.js
@@ -1,0 +1,5 @@
+import React from 'react';
+import { render } from 'react-snapshot';
+import App from './App';
+
+render(<App />, document.getElementById('root'));
