+++ client/webpack/commonConfig.js
@@ -1,0 +1,48 @@
+const path = require('path');
+const webpack = require('webpack');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+
+module.exports = (dir) => ({
+	output: {
+		filename: 'static/js/[name].[contenthash:8].js',
+		assetModuleFilename: 'assets/images/[hash][ext][query]',
+		chunkFilename: 'static/js/[name].[contenthash:8].js',
+	},
+	module: {
+		rules: [
+			{
+				test: /\.(js|jsx)$/,
+				use: ['babel-loader'],
+			},
+			{
+				test: /\.(css|scss)$/,
+				use: [MiniCssExtractPlugin.loader, 'css-loader', 'sass-loader'],
+			},
+			{
+				test: /\.html$/,
+				use: ['html-loader'],
+			},
+			{
+				test: /\.(?:ico|gif|png|jpg|jpeg|svg)$/i,
+				type: 'asset/resource',
+			},
+		],
+	},
+	plugins: [
+		new webpack.ProvidePlugin({
+			Buffer: ['buffer', 'Buffer'],
+		}),
+		new webpack.ProvidePlugin({
+			process: 'process/browser',
+		}),
+		new MiniCssExtractPlugin({
+			filename: 'assets/css/[name].css',
+		}),
+	],
+	resolve: {
+		extensions: ['.js', '.jsx'],
+		alias: {
+			assets: path.resolve(dir, 'src/assets'),
+		},
+	},
+});
