+++ client/webpack/devConfig.js
@@ -1,0 +1,14 @@
+module.exports = (isWeb) =>
+	isWeb
+		? {
+				devServer: {
+					client: {
+						overlay: false,
+					},
+					historyApiFallback: true,
+					port: 3000,
+				},
+		  }
+		: {
+				devtool: 'cheap-module-source-map',
+		  };
