+++ client/webpack/index.js
@@ -1,0 +1,11 @@
+const productionConfig = require('./productionConfig');
+const extensionConfig = require('./extensionConfig');
+const devConfig = require('./devConfig');
+const webConfig = require('./webConfig');
+
+module.exports = {
+	productionConfig,
+	extensionConfig,
+	devConfig,
+	webConfig,
+};
